requirements.txt
google-generativeai>=0.8.0
streamlit==1.28.2
streamlit-chat==0.1.1
python-dotenv==1.0.0
streamlit-extras==0.3.6
Pillow>=10.3.0
torch==2.1.0
